forgot to cange a type of variables in exercise 6

+++ UPPract13.10.21.cpp
@@ -69,7 +69,7 @@ int main()
     cout << divident;
 
     //Задача 6 
-    int side1, side2, side3;
+    double side1, side2, side3;
     cin >> side1 >> side2 >> side3;
     if (side1 + side2 > side3 && side1 + side3 > side2 && side3 + side2 > side1)
     {
